refactor to basic Screen component
